refactor: BaseEntity 디렉토리 이동
