+++ src/main/java/com/springboot/monew/interest/entity/Keyword.java
@@ -1,6 +1,6 @@
 package com.springboot.monew.interest.entity;
 
-import com.springboot.monew.base.BaseEntity;
+import com.springboot.monew.common.entity.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
